Configuração: Removidas frações de minutos

+++ migrations/20150422195606_configuracao.js
@@ -3,10 +3,10 @@
 exports.up = function(knex) {
   return knex.schema.createTable('configuracao', function(table) {
     table.increments('id').primary();
-    table.decimal('franquia_minutos').notNullable();
-    table.decimal('tempo_tolerancia_minutos').notNullable();
-    table.decimal('ciclo_ativacao_minutos').notNullable();
-    table.decimal('ciclo_fiscalizacao_minutos').notNullable();
+    table.integer('franquia_minutos').notNullable();
+    table.integer('tempo_tolerancia_minutos').notNullable();
+    table.integer('ciclo_ativacao_minutos').notNullable();
+    table.integer('ciclo_fiscalizacao_minutos').notNullable();
     table.decimal('valor_ativacao_reais', 18, 2).notNullable();
     table.json('parametros');
     table.integer('cidade_id').notNullable()
